+++ tests/unit/pages/ModelizerDrawPage.spec.js
@@ -1,7 +1,7 @@
 import { installQuasarPlugin } from '@quasar/quasar-app-extension-testing-unit-jest';
 import { shallowMount } from '@vue/test-utils';
 import { Notify } from 'quasar';
-import ModelizerDrawPage from 'src/pages/ModelizerDrawPage.vue';
+import ModelizerDrawPage from 'pages/ModelizerDrawPage.vue';
 import PluginManager from 'src/composables/PluginManager';
 import PluginEvent from 'src/composables/events/PluginEvent';
 import DrawerEvent from 'src/composables/events/DrawerEvent';
@@ -62,7 +62,7 @@ jest.mock('src/composables/PluginManager', () => ({
     resize: jest.fn(),
     exportSvg: jest.fn(() => 'content'),
   })),
-  initComponents: jest.fn(() => Promise.resolve()),
+  initComponents: jest.fn(() => Promise.resolve([])),
   renderConfiguration: jest.fn(() => Promise.resolve()),
   renderModel: jest.fn(() => Promise.resolve()),
   addNewTemplateComponent: jest.fn(() => Promise.resolve()),
@@ -133,6 +133,18 @@ describe('Test page component: ModelizerDrawPage', () => {
     wrapper = shallowMount(ModelizerDrawPage);
   });
 
+  describe('Test function: getStyle', () => {
+    it('should set offset and return valid style', () => {
+      wrapper.vm.offset = 0;
+
+      expect(wrapper.vm.getStyle(100)).toEqual({
+        'min-height': 'calc(100vh - 100px)',
+        'max-height': 'calc(100vh - 100px)',
+      });
+      expect(wrapper.vm.offset).toEqual(100);
+    });
+  });
+
   describe('Test function: onDefaultEvent', () => {
     it('should call renderConfiguration if event.action is "move"', async () => {
       const param = {
@@ -143,7 +155,7 @@ describe('Test page component: ModelizerDrawPage', () => {
 
       await wrapper.vm.onDefaultEvent(param);
 
-      expect(PluginManager.renderConfiguration).toBeCalledWith('project-00000000', 'path', wrapper.vm.data.plugin);
+      expect(PluginManager.renderConfiguration).toBeCalledWith('project-00000000', 'path', wrapper.vm.plugin);
     });
 
     it('should call renderModel if event.action is either "update", "delete" or "add"', async () => {
@@ -155,7 +167,7 @@ describe('Test page component: ModelizerDrawPage', () => {
 
       await wrapper.vm.onDefaultEvent(param);
 
-      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.data.plugin);
+      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.plugin);
 
       param = {
         event: {
@@ -164,7 +176,7 @@ describe('Test page component: ModelizerDrawPage', () => {
       };
       await wrapper.vm.onDefaultEvent(param);
 
-      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.data.plugin);
+      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.plugin);
 
       param = {
         event: {
@@ -173,7 +185,42 @@ describe('Test page component: ModelizerDrawPage', () => {
       };
       await wrapper.vm.onDefaultEvent(param);
 
-      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.data.plugin);
+      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.plugin);
+    });
+  });
+
+  describe('Test function: onDrawerEvent', () => {
+    it('should do nothing on invalid event key', () => {
+      wrapper.vm.splitterKey = null;
+      wrapper.vm.onDrawerEvent({ key: 'ConsoleFooter', type: 'open' });
+      expect(wrapper.vm.splitterKey).toEqual(null);
+    });
+
+    it('should set all variable on valid event', () => {
+      wrapper.vm.componentId = null;
+      wrapper.vm.splitterKey = null;
+      wrapper.vm.isVisible = false;
+      wrapper.vm.splitter = 100;
+      wrapper.vm.onDrawerEvent({ key: 'ComponentDetailPanel', type: 'open', id: 'id_1' });
+      expect(wrapper.vm.componentId).toEqual('id_1');
+      expect(wrapper.vm.splitterKey).toEqual('ComponentDetailPanel');
+      expect(wrapper.vm.isVisible).toEqual(true);
+      expect(wrapper.vm.splitter).toEqual(60);
+
+      wrapper.vm.onDrawerEvent({ key: 'ComponentDetailPanel', type: 'close' });
+      expect(wrapper.vm.componentId).toEqual(null);
+      expect(wrapper.vm.splitterKey).toEqual('ComponentDetailPanel');
+      expect(wrapper.vm.isVisible).toEqual(false);
+      expect(wrapper.vm.splitter).toEqual(100);
+    });
+
+    it('should set next state for AIDrawer on valid event', () => {
+      wrapper.vm.aiDrawerNextState = null;
+      wrapper.vm.onDrawerEvent({ key: 'AIChatDrawer', type: 'open' });
+      expect(wrapper.vm.aiDrawerNextState).toEqual('close');
+
+      wrapper.vm.onDrawerEvent({ key: 'AIChatDrawer', type: 'close' });
+      expect(wrapper.vm.aiDrawerNextState).toEqual('open');
     });
   });
 
@@ -181,14 +228,14 @@ describe('Test page component: ModelizerDrawPage', () => {
     it('should arrange a component position and redraw on click', async () => {
       await wrapper.vm.arrangeComponentsPosition();
 
-      expect(wrapper.vm.data.plugin.arrangeComponentsPosition).toHaveBeenCalled();
-      expect(wrapper.vm.data.plugin.draw).toHaveBeenCalled();
+      expect(wrapper.vm.plugin.arrangeComponentsPosition).toHaveBeenCalled();
+      expect(wrapper.vm.plugin.draw).toHaveBeenCalled();
     });
   });
 
   describe('Test function: initView', () => {
     it('should call draw with "view-port" parameter', async () => {
-      const initComponentsMock = jest.fn(() => Promise.resolve());
+      const initComponentsMock = jest.fn(() => Promise.resolve([]));
 
       PluginManager.initComponents.mockImplementation(initComponentsMock);
       expect(initComponentsMock).toHaveBeenCalledTimes(0);
@@ -196,8 +243,8 @@ describe('Test page component: ModelizerDrawPage', () => {
       await wrapper.vm.initView();
 
       expect(PluginManager.initComponents).toHaveBeenCalled();
-      expect(wrapper.vm.data.plugin.initDrawer).toHaveBeenCalledWith('view-port', false);
-      expect(wrapper.vm.data.plugin.draw).toHaveBeenCalled();
+      expect(wrapper.vm.plugin.initDrawer).toHaveBeenCalledWith('view-port', false);
+      expect(wrapper.vm.plugin.draw).toHaveBeenCalled();
     });
   });
 
@@ -205,11 +252,11 @@ describe('Test page component: ModelizerDrawPage', () => {
     it('should call exportSvg with "view-port" parameter', () => {
       global.URL.createObjectURL = jest.fn();
       global.URL.revokeObjectURL = jest.fn();
-      expect(wrapper.vm.data.plugin.exportSvg).toHaveBeenCalledTimes(0);
+      expect(wrapper.vm.plugin.exportSvg).toHaveBeenCalledTimes(0);
 
       wrapper.vm.exportSvg();
 
-      expect(wrapper.vm.data.plugin.exportSvg).toHaveBeenCalledWith('view-port');
+      expect(wrapper.vm.plugin.exportSvg).toHaveBeenCalledWith('view-port');
     });
   });
 
@@ -227,8 +274,8 @@ describe('Test page component: ModelizerDrawPage', () => {
 
       await wrapper.vm.dropHandler(param);
 
-      expect(wrapper.vm.data.plugin.addComponent).toBeCalled();
-      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.data.plugin);
+      expect(wrapper.vm.plugin.addComponent).toBeCalled();
+      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.plugin);
     });
 
     it('should call addNewTemplateComponent if component is a template and call renderModel', async () => {
@@ -245,7 +292,7 @@ describe('Test page component: ModelizerDrawPage', () => {
       await wrapper.vm.dropHandler(param);
 
       expect(PluginManager.addNewTemplateComponent).toBeCalled();
-      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.data.plugin);
+      expect(PluginManager.renderModel).toBeCalledWith('project-00000000', 'path', wrapper.vm.plugin);
     });
 
     it('should emit a notification on error when adding a new template component', async () => {
@@ -272,37 +319,37 @@ describe('Test page component: ModelizerDrawPage', () => {
     it('should resize and draw', async () => {
       await wrapper.vm.initView();
 
-      wrapper.vm.data.plugin.resize.mockClear();
-      wrapper.vm.data.plugin.draw.mockClear();
+      wrapper.vm.plugin.resize.mockClear();
+      wrapper.vm.plugin.draw.mockClear();
 
       wrapper.vm.onRequestEvent({ type: 'fitToContent', id: 1 });
 
-      expect(wrapper.vm.data.plugin.resize).toBeCalledWith(1);
-      expect(wrapper.vm.data.plugin.draw).toBeCalled();
+      expect(wrapper.vm.plugin.resize).toBeCalledWith(1);
+      expect(wrapper.vm.plugin.draw).toBeCalled();
     });
 
     it('should arrange components position and draw', async () => {
       await wrapper.vm.initView();
 
-      wrapper.vm.data.plugin.arrangeComponentsPosition.mockClear();
-      wrapper.vm.data.plugin.draw.mockClear();
+      wrapper.vm.plugin.arrangeComponentsPosition.mockClear();
+      wrapper.vm.plugin.draw.mockClear();
 
       wrapper.vm.onRequestEvent({ type: 'arrangeContent', id: 1 });
 
-      expect(wrapper.vm.data.plugin.arrangeComponentsPosition).toBeCalledWith(1, false);
-      expect(wrapper.vm.data.plugin.draw).toBeCalled();
+      expect(wrapper.vm.plugin.arrangeComponentsPosition).toBeCalledWith(1, false);
+      expect(wrapper.vm.plugin.draw).toBeCalled();
     });
 
     it('should remove component and draw', async () => {
       await wrapper.vm.initView();
 
-      wrapper.vm.data.plugin.data.removeComponentById.mockClear();
-      wrapper.vm.data.plugin.draw.mockClear();
+      wrapper.vm.plugin.data.removeComponentById.mockClear();
+      wrapper.vm.plugin.draw.mockClear();
 
       wrapper.vm.onRequestEvent({ type: 'delete', id: 1 });
 
-      expect(wrapper.vm.data.plugin.data.removeComponentById).toBeCalledWith(1);
-      expect(wrapper.vm.data.plugin.draw).toBeCalled();
+      expect(wrapper.vm.plugin.data.removeComponentById).toBeCalledWith(1);
+      expect(wrapper.vm.plugin.draw).toBeCalled();
       expect(DrawerEvent.next).toBeCalledWith({ type: 'close', key: 'ComponentDetailPanel' });
     });
 
@@ -314,12 +361,12 @@ describe('Test page component: ModelizerDrawPage', () => {
         setLinkAttribute: jest.fn(),
       };
 
-      wrapper.vm.data.plugin.data.getComponentById = jest.fn((id) => {
+      wrapper.vm.plugin.data.getComponentById = jest.fn((id) => {
         component.id = id;
         return component;
       });
-      wrapper.vm.data.plugin.arrangeComponentsPosition.mockClear();
-      wrapper.vm.data.plugin.draw.mockClear();
+      wrapper.vm.plugin.arrangeComponentsPosition.mockClear();
+      wrapper.vm.plugin.draw.mockClear();
 
       wrapper.vm.onRequestEvent({
         type: 'linkToDefinition',
@@ -332,8 +379,8 @@ describe('Test page component: ModelizerDrawPage', () => {
 
       expect(component.id).toEqual(1);
       expect(component.setLinkAttribute).toBeCalled();
-      expect(wrapper.vm.data.plugin.arrangeComponentsPosition).toBeCalledWith(null, true);
-      expect(wrapper.vm.data.plugin.draw).toBeCalled();
+      expect(wrapper.vm.plugin.arrangeComponentsPosition).toBeCalledWith(null, true);
+      expect(wrapper.vm.plugin.draw).toBeCalled();
     });
 
     it('should create component, setLink and draw', async () => {
@@ -344,11 +391,11 @@ describe('Test page component: ModelizerDrawPage', () => {
         setLinkAttribute: jest.fn(),
       };
 
-      wrapper.vm.data.plugin.data.getComponentById = jest.fn((id) => {
+      wrapper.vm.plugin.data.getComponentById = jest.fn((id) => {
         component.id = id;
         return component;
       });
-      wrapper.vm.data.plugin.draw.mockClear();
+      wrapper.vm.plugin.draw.mockClear();
 
       wrapper.vm.onRequestEvent({
         type: 'linkToComponent',
@@ -361,15 +408,15 @@ describe('Test page component: ModelizerDrawPage', () => {
 
       expect(component.id).toEqual(1);
       expect(component.setLinkAttribute).toBeCalled();
-      expect(wrapper.vm.data.plugin.draw).toBeCalled();
+      expect(wrapper.vm.plugin.draw).toBeCalled();
     });
 
     it('should create component, arrange position and draw', async () => {
       await wrapper.vm.initView();
 
-      wrapper.vm.data.plugin.addComponent.mockClear();
-      wrapper.vm.data.plugin.arrangeComponentsPosition.mockClear();
-      wrapper.vm.data.plugin.draw.mockClear();
+      wrapper.vm.plugin.addComponent.mockClear();
+      wrapper.vm.plugin.arrangeComponentsPosition.mockClear();
+      wrapper.vm.plugin.draw.mockClear();
 
       wrapper.vm.onRequestEvent({
         type: 'addComponentToContainer',
@@ -379,13 +426,13 @@ describe('Test page component: ModelizerDrawPage', () => {
         },
       });
 
-      expect(wrapper.vm.data.plugin.addComponent).toBeCalledWith(
+      expect(wrapper.vm.plugin.addComponent).toBeCalledWith(
         1,
         'definition',
         'project-00000000/path',
       );
-      expect(wrapper.vm.data.plugin.arrangeComponentsPosition).toBeCalledWith(1, true);
-      expect(wrapper.vm.data.plugin.draw).toBeCalled();
+      expect(wrapper.vm.plugin.arrangeComponentsPosition).toBeCalledWith(1, true);
+      expect(wrapper.vm.plugin.draw).toBeCalled();
     });
 
     it('should create component, arrange position and draw', () => {
@@ -410,7 +457,7 @@ describe('Test page component: ModelizerDrawPage', () => {
         type: 'select',
         ids: [1],
       });
-      expect(wrapper.vm.data.plugin.draw).toBeCalled();
+      expect(wrapper.vm.plugin.draw).toBeCalled();
     });
 
     it('should go to editor page and send event to set active file', async () => {
@@ -450,44 +497,6 @@ describe('Test page component: ModelizerDrawPage', () => {
     });
   });
 
-  describe('Test function: setAIDrawerNextState', () => {
-    it('should not change set state on invalid key', () => {
-      DrawerEvent.next.mockClear();
-      wrapper.vm.setAIDrawerNextState({ key: 'invalid', type: 'open' });
-      wrapper.vm.askAI();
-      expect(DrawerEvent.next.mock.calls).toEqual([
-        [{ key: 'ComponentDetailPanel', type: 'close' }],
-        [{ key: 'AIChatDrawer', type: 'open' }],
-      ]);
-
-      DrawerEvent.next.mockClear();
-      wrapper.vm.setAIDrawerNextState({ key: 'invalid', type: 'close' });
-      wrapper.vm.askAI();
-      expect(DrawerEvent.next.mock.calls).toEqual([
-        [{ key: 'ComponentDetailPanel', type: 'close' }],
-        [{ key: 'AIChatDrawer', type: 'open' }],
-      ]);
-    });
-
-    it('should set state', () => {
-      DrawerEvent.next.mockClear();
-      wrapper.vm.setAIDrawerNextState({ key: 'AIChatDrawer', type: 'open' });
-      wrapper.vm.askAI();
-      expect(DrawerEvent.next.mock.calls).toEqual([
-        [{ key: 'ComponentDetailPanel', type: 'close' }],
-        [{ key: 'AIChatDrawer', type: 'close' }],
-      ]);
-
-      DrawerEvent.next.mockClear();
-      wrapper.vm.setAIDrawerNextState({ key: 'AIChatDrawer', type: 'close' });
-      wrapper.vm.askAI();
-      expect(DrawerEvent.next.mock.calls).toEqual([
-        [{ key: 'ComponentDetailPanel', type: 'close' }],
-        [{ key: 'AIChatDrawer', type: 'open' }],
-      ]);
-    });
-  });
-
   describe('Test hook function: onMounted', () => {
     it('should subscribe DefaultEvent', () => {
       expect(pluginDefaultSubscribe).toHaveBeenCalledTimes(1);
@@ -500,7 +509,7 @@ describe('Test page component: ModelizerDrawPage', () => {
   });
 
   describe('Test hook function: onUnmounted', () => {
-    it('should unsubscribe DefaultEvent', () => {
+    it('should unsubscribe PluginEvent', () => {
       expect(pluginDefaultUnsubscribe).toHaveBeenCalledTimes(0);
       expect(pluginRequestUnsubscribe).toHaveBeenCalledTimes(0);
       wrapper.unmount();